ai-assistant/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.0
pytest-html==4.1.1
xmltodict==0.13.0
markdown==3.5.2 